IKBD: read previous mouse packet deltas as signed when merging
MergeMousePacket() read the queued packet's deltas from the unsigned
byte buffer, so a negative delta such as -5 read as 251. Merging two
leftward or upward moves then either failed or, when the sum happened
to be in range, replaced them with a delta in the wrong direction.

+++ src/ikbd.cpp
@@ -417,11 +417,11 @@ void IKBD::MergeMousePacket(int *relx, int *rely, int buttons)
 		return;
 
 	/* Check if distances are not too far */
-	distx = *relx + inbuffer[(mouse_inwrite+1) & (inbufferlen-1)];
+	distx = *relx + (int8)inbuffer[(mouse_inwrite+1) & (inbufferlen-1)];
 	if (abs(distx) > MOUSE_DELTA_MAX)
 		return;
 
-	disty = *rely + inbuffer[(mouse_inwrite+2) & (inbufferlen-1)];
+	disty = *rely + (int8)inbuffer[(mouse_inwrite+2) & (inbufferlen-1)];
 	if (abs(disty) > MOUSE_DELTA_MAX)
 		return;
